requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.3.0  # YOLOv11 (latest)
tensorboard>=2.15.0

# Object Detection Libraries
pycocotools>=2.0.7
albumentations>=1.3.1

# Roboflow
roboflow>=1.1.0

# Visualization
plotly>=5.17.0
tqdm>=4.66.0

# Metrics and Evaluation
torchmetrics>=1.2.0

# Jupyter and Colab
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Utils
PyYAML>=6.0
python-dotenv>=1.0.0
